Envío de Correos Asíncronos con Celery y KeyDB/requirements.txt
Flask==2.3.2
Flask-Mail==0.9.1
celery==5.3.1
redis==4.5.5
python-dotenv==1.0.0
uuid==1.30